Rewrote low-level structure of theories
There was a lot of confusion about expressions and declarations (nodes),
causing redundancy with regard to the type system. Since we have both
built-in types and user-defined types, types now have a separate class.

Nodes have a type, name and possibly definition, which might be an
expression. Statements (they have proofs) and rules (they have even more)
don't follow this structure. (yet?)

Parameters of predicates and rules can now be of any type. This allows
rules for predicate logic, which we couldn't write down before.